smdk4412-common: doze: properly fix future timestamp error
Previously attempted fix at review.lineageos.org/#/c/104083/
The timestamp isn't in the future, that would be impossible.
The real issue is the timestamp from the SensorEvent is a true
timestamp, the System.nanoTime is nanoseconds since boot.
So the previous patch would always return false and never pulse.
I've simply changed it so that the wakeup call gets sent the current
System.nanoTime in milliseconds since boot so it does not think the
event came from the future since the wakeUp method expects
uptimeMillis, not true world time.

linusdan edit: Update gerrit link.

Change-Id: I78f5e9d61107caf9c3bf6c10fa7c5bb400082294